xiaohongshu_automation/requirements.txt
anthropic>=0.18.0
pyyaml>=6.0
schedule>=1.2.0
pillow>=10.0.0
